refactor(block): separate v1/v2 precomputed mods
